Support jruby 1.7.22

+++ run_tests.sh
@@ -19,13 +19,23 @@ killtree() {
 }
 
 # BEGIN TIMEOUT #
-TIMEOUT=3000 # 50 minutes
+TIMEOUT=6000 # 1hour 20 minutes
+PROGRESS_INTERVAL=300 # 5 minutes
 BOSSPID=$$
 (
-  t="/tmp/$$.sh" ; echo 'echo $PPID' > $t
-  BASHPID=`bash $t`
-  rm $t
-  sleep $TIMEOUT
+  if [ "${BASH_VERSINFO[0]}" -lt 4 ] ; then
+    echo "Setting BASHPID for bash < v4"
+    t="/tmp/$$.sh" ; echo 'echo $PPID' > $t
+    BASHPID=`bash $t`
+    rm $t
+  fi
+  if [[ "$TRAVIS" = "true" ]] ; then
+    echo "Wake travis every $PROGRESS_INTERVAL seconds"
+    timeout $TIMEOUT bash -c -- "while true; do sleep $PROGRESS_INTERVAL ; printf '...';done"
+  else
+    echo "Set timeout to $TIMEOUT seconds."
+    sleep $TIMEOUT
+  fi
   echo
   echo "Test timed out after $TIMEOUT seconds."
   echo
